books/serializers.py
from rest_framework import serializers

from .models import Books

class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Books
        fields = ("id", "title", "author", "pages", "description", "price", "created_at")